+++ src/qtviz/elements/bars.py
@@ -6,7 +6,7 @@ from typing import Literal
 
 from ..core.color import ColorSpec
 from ..core.element import Element
-from ..data import Accessor, as_data_ref
+from ..data import Accessor, DataLike, as_data_ref
 
 
 class Bars(Element):
@@ -18,7 +18,7 @@ class Bars(Element):
 
     def __init__(
         self,
-        data,
+        data: DataLike,
         *,
         x: Accessor,
         y: Accessor,